streamlit_app/requirements.txt
streamlit>=1.42.0
marker-pdf>=1.5.5
torch>=2.5.1 